requirements.txt
torch-ema
ninja
trimesh
opencv-python
tensorboardX
numpy 
pandas
tqdm
matplotlib
PyMCubes
rich
dearpygui
packaging
scipy
scikit-learn

face_alignment
python_speech_features
numba
resampy
#pyaudio
soundfile
einops
configargparse

lpips==0.1.3
imageio-ffmpeg

transformers
edge_tts
flask
flask_sockets
opencv-python-headless
aiortc
aiohttp_cors

ffmpeg-python
omegaconf
diffusers
accelerate

librosa
httpx~=0.27.2
